Fix VC++ build error (ulong-->uint_fast64_t)

+++ src/random.cpp
@@ -2,6 +2,7 @@
 
 
 #include <random>
+#include <cstdint>
 #include "random.h"
 namespace nmlib{
 
@@ -11,12 +12,12 @@ static std::uniform_real_distribution<double> URAND(0.0,1.0);
 static std::normal_distribution      <double> NRAND(0.0,1.0);
 static std::exponential_distribution <double> ERAND(1.0);
 
-void init_rand(ulong seed){
+void init_rand(uint_fast64_t seed){
   if(seed==0) seed=std::random_device()();
   MT=std::mt19937_64(seed);
 }
 
-ulong  irand(void){ return ulong(MT()); }
+uint_fast64_t irand(void){ return MT(); }
 double urand(void){ return URAND(MT); }
 double nrand(void){ return NRAND(MT); }
 double erand(void){ return ERAND(MT); }
